Fix test configuration
Fix configuration by not clearing sources, which was causing UseUrls() and any other changes made to the host being lost.
Explicitly set EnterpriseDomain to empty so my local secrets.json changes don't overwrite in in the tests without having to try and inspect the sources and manually remove user secrets.